+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Changed
 
 * `FieldPropsOverride`'s `onChange` now works like schema's and receives [form control context](https://github.com/dgonz64/react-hook-form-auto/blob/master/docs/schema-onchange.md#context).
+* `onChange` is not considered experimental anymore
 
 ### Fixed